refactor: uppercase component names
